Set a maximum of refreshes when wall is empty
The previous behaviour was to try to delete 5000 posts, even if the wall
is empty. The drawback was that if you want to delete only 10 posts,
then you would wait 5000*5 secondes (~7 hours) before the end of the
script.
Now, you can set a number of tries before quitting. If the delete button
is not found consecutively during this number of times, then we break
the loop.

I also added an environment variable parser like argparse to ease the
future use of other env variables.